backend/requirements.txt
fastapi
uvicorn[standard]
python-docx
docx2txt
python-multipart
SQLAlchemy>=2.0
asyncpg
python-dotenv
